feat(tsoa-setup): Switched from legacy to tsoa controller
