+++ client/src/pages/ChangePassword.tsx
@@ -66,7 +66,7 @@ const ChangePassword: React.FunctionComponent = () => {
 
       // Patch request
       api
-        .patch("/users/change-password", data, {
+        .patch("/members/change-password", data, {
           headers: { Authorization: `Bearer ${localStorage.getItem("token")}` },
         })
         .then((res) => {
