+++ client/src/pages/members/MemberOverview.tsx
@@ -159,7 +159,7 @@ const MemberOverview: React.FunctionComponent = () => {
     // Variable for checking, if the component is mounted
     let mounted = true;
     api
-      .get("/users/", {
+      .get("/members/", {
         headers: { Authorization: `Bearer ${localStorage.getItem("token")}` },
       })
       .then((res) => {
