+++ server/tsoa.json
@@ -14,6 +14,7 @@
     }
   },
   "routes": {
+    "basePath": "/api",
     "routesDir": "build",
     "authenticationModule": "src/middleware/authenticationTsoa.ts"
   }
